% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{between_class_neighbors}
\alias{between_class_neighbors}
\title{Between-Class Neighbors}
\usage{
between_class_neighbors(x, ng, ...)
}
\arguments{
\item{x}{An object.}

\item{ng}{A neighbor graph object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
An object representing the between-class neighbors of the input graph, the structure of which depends on the input object's class.
}
\description{
A generic function to compute the between-class neighbors of a graph.
}
\examples{
labs <- factor(c("a","a","b"))
cg <- class_graph(labs)
ng <- neighbor_graph(matrix(c(0,1,1,1,0,1,1,1,0),3))
between_class_neighbors(cg, ng)

}
