% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{find_nn}
\alias{find_nn}
\title{Find nearest neighbors}
\usage{
find_nn(x, ...)
}
\arguments{
\item{x}{An object of class "nnsearcher".}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A nearest neighbors result object.
}
\description{
Find nearest neighbors
}
\examples{
\donttest{
X <- matrix(rnorm(20), nrow=5)
nn <- nnsearcher(X)
find_nn(nn, k=2)
}
}
