% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{find_nn_between}
\alias{find_nn_between}
\title{Find nearest neighbors between two sets of data points}
\usage{
find_nn_between(x, ...)
}
\arguments{
\item{x}{An object of class "nnsearcher".}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A nearest neighbors result object.
}
\description{
Find nearest neighbors between two sets of data points
}
\examples{
\donttest{
X <- matrix(rnorm(20), nrow=5)
nn <- nnsearcher(X)
find_nn_between(nn, k=1, idx1=1:2, idx2=3:5)
}
}
