% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{inverse_heat_kernel}
\alias{inverse_heat_kernel}
\title{inverse_heat_kernel}
\usage{
inverse_heat_kernel(x, sigma = 1)
}
\arguments{
\item{x}{the distances}

\item{sigma}{the bandwidth}
}
\value{
Numeric vector/matrix of inverse heat kernel values.
}
\description{
inverse_heat_kernel
}
\examples{
inverse_heat_kernel(c(1, 2), sigma = 1)
}
