% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{non_neighbors}
\alias{non_neighbors}
\title{Get Indices of Non-neighbors of a Node}
\usage{
non_neighbors(x, ...)
}
\arguments{
\item{x}{A neighbor graph object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A numeric vector of node indices that are not neighbors of the given node.
}
\description{
Retrieve the indices of nodes that are not neighbors of a specified node.
}
\examples{
adj_matrix <- matrix(c(0,1,0,
                       1,0,0,
                       0,0,0), nrow=3, byrow=TRUE)
ng <- neighbor_graph(adj_matrix)
non_neighbors(ng, 1)
}
