% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{search_result.nnsearcher}
\alias{search_result.nnsearcher}
\title{Convert Search Results for nnsearcher Objects}
\usage{
\method{search_result}{nnsearcher}(x, result)
}
\arguments{
\item{x}{An object of class "nnsearcher".}

\item{result}{A raw result object from nearest neighbor search.}
}
\value{
An object of class "nn_search" with standardized field names.
}
\description{
Convert raw search results to a standardized format for nnsearcher objects.
}
\examples{
res <- list(idx = matrix(c(1L,2L), nrow=1),
            dist = matrix(c(0.1,0.2), nrow=1))
dummy <- nnsearcher(matrix(rnorm(4), nrow=2))
search_result(dummy, res)
}
