% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{search_result}
\alias{search_result}
\title{Search result for nearest neighbor search}
\usage{
search_result(x, result)
}
\arguments{
\item{x}{An object of class "nnsearcher".}

\item{result}{The result from the nearest neighbor search.}
}
\value{
An object with the class "nn_search".
}
\description{
Search result for nearest neighbor search
}
\examples{
res <- list(idx = matrix(c(1L,2L), nrow=1),
            dist = matrix(c(0.1,0.2), nrow=1))
dummy <- nnsearcher(matrix(rnorm(4), nrow=2))
search_result(dummy, res)
}
